+++ tests/train_test.py
@@ -6,6 +6,7 @@ import pandas as pd
 
 
 def test_train_model_mexican():
+    ###Test for positive pairs datasets
     # Define the path to the test dataset
     dataset_path = os.path.join(DATA_DIR_PATH, "es_mexican_products.xlsx")
 
@@ -29,14 +30,14 @@ def test_train_model_mexican():
 
     model=lt.load_model(saved_model_path)
 
-    model.save_to_hub(repo_name = "linktransformer-models-test", ##Write model name here
-                    organization= "dell-research-harvard",
-                    private = None,
-                    commit_message = "Add new LinkTransformer model.",
-                    local_model_path = None,
-                    exist_ok = True,
-                    replace_model_card = True,
-                    )
+    # model.save_to_hub(repo_name = "linktransformer-models-test", ##Write model name here
+    #                 organization= "dell-research-harvard",
+    #                 private = None,
+    #                 commit_message = "Add new LinkTransformer model.",
+    #                 local_model_path = None,
+    #                 exist_ok = True,
+    #                 replace_model_card = True,
+    #                 )
 
     # Add assertions to check if the training was successful and the model was saved
     assert os.path.exists(saved_model_path), "Model not saved"
@@ -44,6 +45,7 @@ def test_train_model_mexican():
 
 
 def test_train_model_jp():
+    ##Test for positive and neg pairs datasets
     dataset_path = os.path.join(DATA_DIR_PATH, "jp_pr_tk431.csv")
 
     ##Load the data
@@ -73,8 +75,8 @@ def test_train_model_jp():
         label_col_name="tk_truth",
         left_id_name=['source'],
         right_id_name=['tk_path_value'],
-        log_wandb=True,
-        training_args={"num_epochs": 0,
+        log_wandb=False,
+        training_args={"num_epochs": 1,
                        "test_at_end": True,
                        "save_val_test_pickles": True,
                        "model_save_name": "check",
@@ -84,7 +86,31 @@ def test_train_model_jp():
     )
 
 
+def test_train_model_clustering():
+    #Test for cluster datasets
+    dataset_path = os.path.join(DATA_DIR_PATH, "/mnt/122a7683-fa4b-45dd-9f13-b18cc4f4a187/deeprecordlinkage/linktransformer/src/linktransformer/data/company_clusters.csv")
+    ##Load the data
+    df = pd.read_csv(dataset_path)
 
-if __name__ == "__main__":
-    # test_train_model_jp()
-    test_train_model_mexican()
+    saved_model_path = lt.train_model(
+        model_path="sentence-transformers/all-mpnet-base-v2",
+        data=df,
+        clus_id_col_name=["cluster_id"],
+        clus_text_col_names=["company_name"],
+        log_wandb=True,
+        training_args={"num_epochs": 1,
+                          "test_at_end": True,
+                            "save_val_test_pickles": True,
+                            "model_save_name": "check2",
+                            "opt_model_description": "test",
+                            "opt_model_lang":"en",
+                            "val_perc":0.2,
+                            "batch_size": 128,
+                            "large_val": True}
+
+    )
+
+
+# if __name__ == "__main__":
+#     # test_train_model_jp()
+#     # test_train_model_mexican()